python/src/auth/pyproject.toml
[tool.poetry]
name = "authentication service"
version = "0.1.0"
description = ""
authors = ["mohamedayari1 <ayarim781@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyjwt = "^2.9.0"
flask = "^3.0.3"
flask-mysqldb = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
